requirements.txt
Authlib==0.14.3
bcrypt==3.2.0
blinker==1.4
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
click==7.1.2
cryptography==3.1
dnspython==2.0.0
email-validator==1.1.1
Flask==1.1.2
Flask-Bcrypt==0.7.1
Flask-Dance==3.0.0
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
mysqlclient==2.0.1
oauthlib==3.1.0
Pillow==7.2.0
pycparser==2.20
requests==2.24.0
requests-oauthlib==1.3.0
six==1.15.0
SQLAlchemy==1.3.19
urllib3==1.25.10
URLObject==2.4.3
Werkzeug==1.0.1
WTForms==2.3.3
